After interview

+++ src/main/java/com/moon/senla/educational_website/service/impl/CourseServiceImpl.java
@@ -1,51 +1,58 @@
 package com.moon.senla.educational_website.service.impl;
 
 
+import static com.moon.senla.educational_website.utils.StringConstants.ACCESS_DENIED;
+import static com.moon.senla.educational_website.utils.StringConstants.COULD_NOT_DELETE;
+import static com.moon.senla.educational_website.utils.StringConstants.COULD_NOT_SAVED;
+import static com.moon.senla.educational_website.utils.StringConstants.COURSE_NF;
+
 import com.moon.senla.educational_website.dao.CourseRepository;
-import com.moon.senla.educational_website.dao.UserRepository;
+import com.moon.senla.educational_website.error.AuthException;
 import com.moon.senla.educational_website.error.CustomException;
+import com.moon.senla.educational_website.error.NotFoundException;
+import com.moon.senla.educational_website.error.ValidationException;
 import com.moon.senla.educational_website.model.Course;
 import com.moon.senla.educational_website.model.User;
-import com.moon.senla.educational_website.model.dto.course.CourseNewDto;
-import com.moon.senla.educational_website.model.dto.course.CourseUpdateDto;
-import com.moon.senla.educational_website.model.dto.mapper.CourseMapper;
 import com.moon.senla.educational_website.service.CourseService;
+import com.moon.senla.educational_website.service.TopicService;
+import com.moon.senla.educational_website.service.UserService;
 import java.security.Principal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 @Service
 public class CourseServiceImpl implements CourseService {
 
     private final CourseRepository courseRepository;
-    private final UserRepository userRepository;
+    private final UserService userService;
+    private final TopicService topicService;
 
     @Autowired
     public CourseServiceImpl(CourseRepository courseRepository,
-        UserRepository userRepository) {
+        UserService userService,
+        TopicService topicService) {
         this.courseRepository = courseRepository;
-        this.userRepository = userRepository;
+        this.userService = userService;
+        this.topicService = topicService;
     }
 
     @Override
-    public Course save(Principal principal, CourseNewDto newCourse) {
+    public Course save(Principal principal, Course course) {
         try {
-            Course course = CourseMapper.INSTANCE.courseNewDtoToCourse(newCourse);
-            course.setUser(userRepository.findByUsername(principal.getName()));
+            course.setUser(userService.findByUsername(principal.getName()));
+            course.setTopic(topicService.findById(course.getTopic().getId()));
             return courseRepository.save(course);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, course could not be saved");
+                throw new ValidationException(COULD_NOT_SAVED.value);
         }
     }
 
     @Override
     public Course findById(long id) {
         return courseRepository.findById(id)
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Course Not Found"));
+            .orElseThrow(() -> new NotFoundException(COURSE_NF.value));
     }
 
     @Override
@@ -53,40 +60,36 @@ public class CourseServiceImpl implements CourseService {
         try {
             return courseRepository.findAll(pageable);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, courses cannot be found");
+            throw new NotFoundException(COURSE_NF.value);
         }
     }
 
     @Override
     public void deleteById(long id) {
-        courseRepository.findById(id)
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Course Not Found"));
+        if (!courseRepository.findById(id).isPresent()) {
+            throw new NotFoundException(COURSE_NF.value);
+        }
         try {
             courseRepository.deleteById(id);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, failed to delete");
+            throw new ValidationException(COULD_NOT_DELETE.value);
         }
     }
 
     @Override
-    public Course update(Principal principal, CourseUpdateDto courseToUpdate) {
+    public Course update(Principal principal, Course courseToUpdate) {
         Course oldCourse = courseRepository.findById(courseToUpdate.getId())
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Course Not Found"));
-        User user = userRepository.findById(oldCourse.getUser().getId())
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "User Not Found"));
+            .orElseThrow(() -> new NotFoundException(COURSE_NF.value));
+        User user = userService.findById(oldCourse.getUser().getId());
         if (!user.getUsername().equals(principal.getName())) {
-            throw new CustomException(HttpStatus.FORBIDDEN,
-                "Invalid request, access is denied");
+            throw new AuthException(ACCESS_DENIED.value);
         }
         oldCourse.setName(courseToUpdate.getName());
         oldCourse.setPrice(courseToUpdate.getPrice());
         try {
             return courseRepository.save(oldCourse);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, course could not be updated");
+            throw new CustomException(ACCESS_DENIED.value);
         }
     }
 
@@ -95,8 +98,17 @@ public class CourseServiceImpl implements CourseService {
         try {
             return courseRepository.findAllCoursesByUsername(pageable, username);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, courses cannot be found");
+            throw new NotFoundException(COURSE_NF.value);
+        }
+    }
+
+    @Override
+    public Page<Course> findAllCoursesByParam(Pageable pageable, String name, String topicName,
+        String authorName) {
+        try {
+            return courseRepository.findAllCoursesByParam(pageable, name, topicName, authorName);
+        } catch (Exception e) {
+            throw new NotFoundException(COURSE_NF.value);
         }
     }
