+++ src/main/java/com/moon/senla/educational_website/service/impl/ScheduleServiceImpl.java
@@ -1,63 +1,65 @@
 package com.moon.senla.educational_website.service.impl;
 
-import com.moon.senla.educational_website.dao.CourseRepository;
-import com.moon.senla.educational_website.dao.GroupRepository;
+import static com.moon.senla.educational_website.utils.StringConstants.ACCESS_DENIED;
+import static com.moon.senla.educational_website.utils.StringConstants.COULD_NOT_DELETE;
+import static com.moon.senla.educational_website.utils.StringConstants.COULD_NOT_SAVED;
+import static com.moon.senla.educational_website.utils.StringConstants.COULD_NOT_UPDATED;
+import static com.moon.senla.educational_website.utils.StringConstants.SCHEDULE_NF;
+
 import com.moon.senla.educational_website.dao.ScheduleRepository;
-import com.moon.senla.educational_website.dao.UserRepository;
-import com.moon.senla.educational_website.error.CustomException;
+import com.moon.senla.educational_website.error.AuthException;
+import com.moon.senla.educational_website.error.NotFoundException;
+import com.moon.senla.educational_website.error.ValidationException;
 import com.moon.senla.educational_website.model.Course;
 import com.moon.senla.educational_website.model.Group;
 import com.moon.senla.educational_website.model.Schedule;
 import com.moon.senla.educational_website.model.User;
-import com.moon.senla.educational_website.model.dto.mapper.ScheduleMapper;
-import com.moon.senla.educational_website.model.dto.schedule.ScheduleNewDto;
-import com.moon.senla.educational_website.model.dto.schedule.ScheduleUpdateDto;
+import com.moon.senla.educational_website.service.CourseService;
+import com.moon.senla.educational_website.service.GroupService;
 import com.moon.senla.educational_website.service.ScheduleService;
+import com.moon.senla.educational_website.service.UserService;
 import java.security.Principal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 @Service
 public class ScheduleServiceImpl implements ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
-    private final GroupRepository groupRepository;
-    private final CourseRepository courseRepository;
-    private final UserRepository userRepository;
+    private final GroupService groupService;
+    private final CourseService courseService;
+    private final UserService userService;
 
     @Autowired
     public ScheduleServiceImpl(
         ScheduleRepository scheduleRepository,
-        GroupRepository groupRepository,
-        CourseRepository courseRepository,
-        UserRepository userRepository) {
+        GroupService groupService,
+        CourseService courseService,
+        UserService userService) {
         this.scheduleRepository = scheduleRepository;
-        this.groupRepository = groupRepository;
-        this.courseRepository = courseRepository;
-        this.userRepository = userRepository;
+        this.groupService = groupService;
+        this.courseService = courseService;
+        this.userService = userService;
     }
 
     @Override
-    public Schedule save(Principal principal, ScheduleNewDto schedule) {
+    public Schedule save(Principal principal, Schedule schedule) {
         Group group = checkRequest(principal, schedule.getGroup().getId());
-        Schedule newSchedule = ScheduleMapper.INSTANCE.scheduleNewDtoToSchedule(schedule);
-        newSchedule.setGroup(group);
-        newSchedule.setDate(schedule.getDate());
+        schedule.setGroup(group);
+        schedule.setDate(schedule.getDate());
         try {
-            return scheduleRepository.save(newSchedule);
+            return scheduleRepository.save(schedule);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, schedule could not be saved");
+            throw new ValidationException(COULD_NOT_SAVED.value);
         }
     }
 
     @Override
     public Schedule findById(long id) {
         return scheduleRepository.findById(id)
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Schedule Not Found"));
+            .orElseThrow(() -> new NotFoundException(SCHEDULE_NF.value));
     }
 
     @Override
@@ -65,59 +67,52 @@ public class ScheduleServiceImpl implements ScheduleService {
         try {
             return scheduleRepository.findAll(pageable);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, schedules cannot be found");
+            throw new NotFoundException(SCHEDULE_NF.value);
         }
     }
 
     @Override
     public void deleteById(Principal principal, long id) {
-        Schedule oldSchedule = scheduleRepository.findById(id)
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Schedule Not Found"));
-        checkRequest(principal, oldSchedule.getGroup().getId());
+        if (!scheduleRepository.findById(id).isPresent()) {
+            throw new NotFoundException(SCHEDULE_NF.value);
+        }
+        checkRequest(principal, id);
         try {
             scheduleRepository.deleteById(id);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.NOT_FOUND, "Schedule Not Found");
+            throw new ValidationException(COULD_NOT_DELETE.value);
         }
     }
 
     @Override
-    public Page<Schedule> findAllByGroup_Id(Pageable pageable, long groupId) {
-        groupRepository.findById(groupId)
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Group Not Found"));
+    public Page<Schedule> findAllByGroupId(Pageable pageable, long groupId) {
+        Group group = groupService.findById(groupId);
         try {
-            return scheduleRepository.findAllByGroup_Id(pageable, groupId);
+            return scheduleRepository.findAllByGroupId(pageable, group.getId());
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, schedules cannot be found");
+            throw new NotFoundException(SCHEDULE_NF.value);
         }
     }
 
     @Override
-    public Schedule update(Principal principal, ScheduleUpdateDto schedule) {
+    public Schedule update(Principal principal, Schedule schedule) {
         Schedule oldSchedule = scheduleRepository.findById(schedule.getId())
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Schedule Not Found"));
+            .orElseThrow(() -> new NotFoundException(SCHEDULE_NF.value));
         checkRequest(principal, oldSchedule.getGroup().getId());
         oldSchedule.setDate(schedule.getDate());
         try {
             return scheduleRepository.save(oldSchedule);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, schedule could not be updated");
+            throw new ValidationException(COULD_NOT_UPDATED.value);
         }
     }
 
     private Group checkRequest(Principal principal, Long id) {
-        Group group = groupRepository.findById(id)
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Group Not Found"));
-        Course course = courseRepository.findById(group.getCourse().getId())
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Course Not Found"));
-        User user = userRepository.findById(course.getUser().getId())
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "User Not Found"));
+        Group group = groupService.findById(id);
+        Course course = courseService.findById(group.getCourse().getId());
+        User user = userService.findById(course.getUser().getId());
         if (!user.getUsername().equals(principal.getName())) {
-            throw new CustomException(HttpStatus.FORBIDDEN,
-                "Invalid request, access is denied");
+            throw new AuthException(ACCESS_DENIED.value);
         }
         return group;
     }
