+++ src/main/java/com/moon/senla/educational_website/service/impl/TopicServiceImpl.java
@@ -1,17 +1,19 @@
 package com.moon.senla.educational_website.service.impl;
 
 
+import static com.moon.senla.educational_website.utils.StringConstants.COULD_NOT_DELETE;
+import static com.moon.senla.educational_website.utils.StringConstants.COULD_NOT_SAVED;
+import static com.moon.senla.educational_website.utils.StringConstants.COULD_NOT_UPDATED;
+import static com.moon.senla.educational_website.utils.StringConstants.TOPIC_NF;
+
 import com.moon.senla.educational_website.dao.TopicRepository;
-import com.moon.senla.educational_website.error.CustomException;
+import com.moon.senla.educational_website.error.NotFoundException;
+import com.moon.senla.educational_website.error.ValidationException;
 import com.moon.senla.educational_website.model.Topic;
-import com.moon.senla.educational_website.model.dto.mapper.TopicMapper;
-import com.moon.senla.educational_website.model.dto.topic.TopicDto;
-import com.moon.senla.educational_website.model.dto.topic.TopicNewDto;
 import com.moon.senla.educational_website.service.TopicService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -25,19 +27,18 @@ public class TopicServiceImpl implements TopicService {
     }
 
     @Override
-    public Topic save(TopicNewDto topic) {
+    public Topic save(Topic topic) {
         try {
-            return topicRepository.save(TopicMapper.INSTANCE.topicNewDtoToTopic(topic));
+            return topicRepository.save(topic);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, topic could not be saved");
+            throw new ValidationException(COULD_NOT_SAVED.value);
         }
     }
 
     @Override
     public Topic findById(long id) {
         return topicRepository.findById(id)
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Topic Not Found"));
+            .orElseThrow(() -> new NotFoundException(TOPIC_NF.value));
     }
 
     @Override
@@ -45,33 +46,31 @@ public class TopicServiceImpl implements TopicService {
         try {
             return topicRepository.findAll(pageable);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, topic cannot be found");
+            throw new NotFoundException(TOPIC_NF.value);
         }
     }
 
     @Override
     public void deleteById(long id) {
-        topicRepository.findById(id)
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Topic Not Found"));
+        if (!topicRepository.findById(id).isPresent()) {
+            throw new NotFoundException(TOPIC_NF.value);
+        }
         try {
             topicRepository.deleteById(id);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, failed to delete");
+            throw new ValidationException(COULD_NOT_DELETE.value);
         }
     }
 
     @Override
-    public Topic update(TopicDto topicUpdate) {
+    public Topic update(Topic topicUpdate) {
         Topic oldTopic = topicRepository.findById(topicUpdate.getId())
-            .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "Topic Not Found"));
+            .orElseThrow(() -> new NotFoundException(TOPIC_NF.value));
         oldTopic.setName(topicUpdate.getName());
         try {
             return topicRepository.save(oldTopic);
         } catch (Exception e) {
-            throw new CustomException(HttpStatus.BAD_REQUEST,
-                "Invalid request, topic could not be updated");
+            throw new ValidationException(COULD_NOT_UPDATED.value);
         }
     }
 }
